+++ apm-sniffer/apm-sdk-plugin/httpasyncclient-4.x-plugin/src/main/java/org/apache/skywalking/apm/plugin/httpasyncclient/v4/HttpAsyncRequestExecutorInterceptor.java
@@ -121,10 +121,9 @@ public class HttpAsyncRequestExecutorInterceptor implements InstanceMethodsAroun
         ContextSnapshot snapshot = (ContextSnapshot) contextInConn.getAttribute(Constants.SKYWALKING_CONTEXT_SNAPSHOT);
         conn.getContext().removeAttribute(Constants.SKYWALKING_CONTEXT_SNAPSHOT);
         if (snapshot != null) {
-            AbstractSpan localSpan = ContextManager.createLocalSpan("HttpAsyncClient/local");
+            AbstractSpan localSpan = ContextManager.createLocalSpan("HttpAsyncClient/local", snapshot);
             localSpan.setComponent(ComponentsDefine.HTTP_ASYNC_CLIENT);
             localSpan.setLayer(SpanLayer.HTTP);
-            ContextManager.continued(snapshot);
         }
         return context;
     }