+++ apm-sniffer/apm-sdk-plugin/httpasyncclient-4.x-plugin/src/main/java/org/apache/skywalking/apm/plugin/httpasyncclient/v4/SessionRequestCompleteInterceptor.java
@@ -42,12 +42,9 @@ public class SessionRequestCompleteInterceptor implements InstanceMethodsAroundI
             return;
         }
         ContextSnapshot snapshot = (ContextSnapshot) array[0];
-        AbstractSpan localSpan = ContextManager.createLocalSpan("httpasyncclient/local");
+        AbstractSpan localSpan = ContextManager.createLocalSpan("httpasyncclient/local", snapshot);
         localSpan.setComponent(ComponentsDefine.HTTP_ASYNC_CLIENT);
         localSpan.setLayer(SpanLayer.HTTP);
-        if (snapshot != null) {
-            ContextManager.continued(snapshot);
-        }
         Constants.HTTP_CONTEXT_LOCAL.set((HttpContext) array[1]);
 
     }