+++ apm-sniffer/apm-sdk-plugin/httpclient-5.x-plugin/src/main/java/org/apache/skywalking/apm/plugin/httpclient/v5/IOSessionImplPollInterceptor.java
@@ -59,10 +59,9 @@ public class IOSessionImplPollInterceptor implements InstanceMethodsAroundInterc
             return ret;
         }
         httpContext.removeAttribute(Constants.SKYWALKING_CONTEXT_SNAPSHOT);
-        AbstractSpan localSpan = ContextManager.createLocalSpan("httpasyncclient/local");
+        AbstractSpan localSpan = ContextManager.createLocalSpan("httpasyncclient/local", snapshot);
         localSpan.setComponent(ComponentsDefine.HTTP_ASYNC_CLIENT);
         localSpan.setLayer(SpanLayer.HTTP);
-        ContextManager.continued(snapshot);
 
         final ContextCarrier contextCarrier = new ContextCarrier();
         BasicHttpRequest request = (BasicHttpRequest) httpContext.getAttribute(HttpClientContext.HTTP_REQUEST);