+++ apm-sniffer/apm-toolkit-activation/apm-toolkit-trace-activation/src/main/java/org/apache/skywalking/apm/toolkit/activation/trace/CallableOrRunnableInvokeInterceptor.java
@@ -29,11 +29,10 @@ public class CallableOrRunnableInvokeInterceptor implements InstanceMethodsAroun
     @Override
     public void beforeMethod(EnhancedInstance objInst, Method method, Object[] allArguments, Class<?>[] argumentsTypes,
         MethodInterceptResult result) throws Throwable {
-        ContextManager.createLocalSpan("Thread/" + objInst.getClass().getName() + "/" + method.getName());
         ContextSnapshot cachedObjects = (ContextSnapshot) objInst.getSkyWalkingDynamicField();
-        if (cachedObjects != null) {
-            ContextManager.continued(cachedObjects);
-        }
+
+        ContextManager.createLocalSpan("Thread/" + objInst.getClass().getName() + "/" + method.getName(), cachedObjects);
+
     }
 
     @Override