+++ apm-sniffer/optional-plugins/optional-spring-plugins/optional-spring-cloud/gateway-2.0.x-plugin/src/main/java/org/apache/skywalking/apm/plugin/spring/cloud/gateway/v20x/NettyRoutingFilterInterceptor.java
@@ -18,6 +18,7 @@
 package org.apache.skywalking.apm.plugin.spring.cloud.gateway.v20x;
 
 import java.lang.reflect.Method;
+
 import org.apache.skywalking.apm.agent.core.context.ContextManager;
 import org.apache.skywalking.apm.agent.core.context.ContextSnapshot;
 import org.apache.skywalking.apm.agent.core.context.trace.AbstractSpan;
@@ -37,10 +38,13 @@ public class NettyRoutingFilterInterceptor implements InstanceMethodsAroundInter
         ServerWebExchange exchange = (ServerWebExchange) allArguments[0];
         EnhancedInstance enhancedInstance = getInstance(exchange);
 
-        AbstractSpan span = ContextManager.createLocalSpan("SpringCloudGateway/RoutingFilter");
+        ContextSnapshot contextSnapshot = null;
         if (enhancedInstance != null && enhancedInstance.getSkyWalkingDynamicField() != null) {
-            ContextManager.continued((ContextSnapshot) enhancedInstance.getSkyWalkingDynamicField());
+            contextSnapshot = (ContextSnapshot) enhancedInstance.getSkyWalkingDynamicField();
         }
+
+        AbstractSpan span = ContextManager.createLocalSpan("SpringCloudGateway/RoutingFilter", contextSnapshot);
+
         span.setComponent(SPRING_CLOUD_GATEWAY);
         span.prepareForAsync();
     }