+++ apm-sniffer/optional-plugins/optional-spring-plugins/optional-spring-cloud/gateway-2.1.x-plugin/src/main/java/org/apache/skywalking/apm/plugin/spring/cloud/gateway/v21x/NettyRoutingFilterInterceptor.java
@@ -45,10 +45,13 @@ public class NettyRoutingFilterInterceptor implements InstanceMethodsAroundInter
 
         EnhancedInstance enhancedInstance = getInstance(exchange);
 
-        AbstractSpan span = ContextManager.createLocalSpan("SpringCloudGateway/RoutingFilter");
+        ContextSnapshot contextSnapshot = null;
         if (enhancedInstance != null && enhancedInstance.getSkyWalkingDynamicField() != null) {
-            ContextManager.continued((ContextSnapshot) enhancedInstance.getSkyWalkingDynamicField());
+            contextSnapshot = (ContextSnapshot) enhancedInstance.getSkyWalkingDynamicField();
         }
+
+        AbstractSpan span = ContextManager.createLocalSpan("SpringCloudGateway/RoutingFilter", contextSnapshot);
+
         span.setComponent(SPRING_CLOUD_GATEWAY);
     }